Sorted into folders
I haven't change any code, so right now its not functional in its current state. I'll start fixing the code for the links on the pages later, but I feel pretty confident about the structure of the directories. I'm a little tipsy right now, so I'll save the referencing for another time